Checks Policy Enforcement Guard
