+++ build.sbt
@@ -11,11 +11,11 @@ lazy val pluginSettings = Seq(
   addSbtPlugin("com.typesafe.sbt"   % "sbt-ghpages"            % "0.6.0"),
   addSbtPlugin("com.typesafe.sbt"   % "sbt-site"               % "1.2.0"),
   addSbtPlugin("org.tpolecat"       % "tut-plugin"             % "0.4.8"),
-  addSbtPlugin("pl.project13.scala" % "sbt-jmh"                % "0.2.23"),
+  addSbtPlugin("pl.project13.scala" % "sbt-jmh"                % "0.2.24"),
   addSbtPlugin("org.scalastyle"     %% "scalastyle-sbt-plugin" % "0.8.0"),
   addSbtPlugin("org.scoverage"      % "sbt-scoverage"          % "1.5.0"),
   addSbtPlugin("com.typesafe.sbt"   % "sbt-git"                % "0.8.5"),
-  addSbtPlugin("org.scala-js"       % "sbt-scalajs"            % "0.6.14"),
+  addSbtPlugin("org.scala-js"       % "sbt-scalajs"            % "0.6.15"),
   addSbtPlugin("com.fortysevendeg"  % "sbt-microsites"         % "0.4.0"),
   addSbtPlugin("de.heikoseeberger"  % "sbt-header"             % "1.8.0"),
   addSbtPlugin("com.eed3si9n"       % "sbt-buildinfo"          % "0.6.1"),
